small security fixes

+++ app/Http/Controllers/SecurityController.php
@@ -21,13 +21,21 @@ class SecurityController extends Controller
     }
     public function deleteProduct($id)
     {
-        $product = Product::find($id);
-        $product->delete();
-        return redirect()->route('security.products');
+        if (Session::get('username') == env("LOGIN_USERNAME") && Session::get('isLoggedIn')) {
+            $product = Product::find($id);
+            $product->delete();
+            return redirect()->route('security.products');
+        } else {
+            echo trans('messages.You must be logged in');
+        }
     }
     public function updateProduct($id)
     {
-        return view('user.product' , ['id' => $id]);
+        if (Session::get('username') == env("LOGIN_USERNAME") && Session::get('isLoggedIn')) {
+            return view('user.product' , ['id' => $id]);
+        } else {
+            echo trans('messages.You must be logged in');
+        }
     }
     public function getProduct()
     {
@@ -39,36 +47,40 @@ class SecurityController extends Controller
     }
     public function addProduct()
     {
-        $addphoto = false;
-        if ($_FILES["fileToUpload"]["size"] != 0) {
-            if (strpos($_FILES["fileToUpload"]["type"], "image") !== false) {
-                $input = $_FILES["fileToUpload"]["tmp_name"];
-                $addphoto = true;
+        if (Session::get('username') == env("LOGIN_USERNAME") && Session::get('isLoggedIn')) {
+            $addphoto = false;
+            if ($_FILES["fileToUpload"]["size"] != 0) {
+                if (strpos($_FILES["fileToUpload"]["type"], "image") !== false) {
+                    $input = $_FILES["fileToUpload"]["tmp_name"];
+                    $addphoto = true;
+                } else {
+                    echo trans('messages.The type of your file is not accepted. We accept image file.');
+                }
             } else {
-                echo trans('messages.The type of your file is not accepted. We accept image file.');
+                echo trans('messages.You did not insert the picture');
             }
-        } else {
-            echo trans('messages.You did not insert the picture');
-        }
-        if($addphoto) {
-            if($_POST['id']){
-                $output = "public/photo/photo-". $_POST["id"] .'.jpg';
-                $product = Product::find($_POST['id']);
-                $product->title = $_POST['Title'];
-                $product->description = $_POST['Description'];
-                $product->price = $_POST['Price'];
-                $product->save();
+            if($addphoto) {
+                if($_POST['id']){
+                    $output = "public/photo/photo-". $_POST["id"] .'.jpg';
+                    $product = Product::find($_POST['id']);
+                    $product->title = $_POST['Title'];
+                    $product->description = $_POST['Description'];
+                    $product->price = $_POST['Price'];
+                    $product->save();
+                } else {
+                    $product = new Product([
+                        'title' => $_POST['Title'],
+                        'description' => $_POST['Description'],
+                        'price' => $_POST['Price']
+                    ]);
+                    $product->save();
+                    $output = "public/photo/photo-".$product->id.'.jpg';
+                }
+                move_uploaded_file(file_get_contents($input),$output);
+                return redirect()->route('security.products');
             } else {
-                $product = new Product([
-                    'title' => $_POST['Title'],
-                    'description' => $_POST['Description'],
-                    'price' => $_POST['Price']
-                ]);
-                $product->save();
-                $output = "public/photo/photo-".$product->id.'.jpg';
+                echo trans('messages.You must be logged in');
             }
-            move_uploaded_file(file_get_contents($input),$output);
-            return redirect()->route('security.products');
         }               
     }
 }
